requirements.txt
# Testing Frameworks
pytest==7.4.3
pytest-html==4.0.2
pytest-xdist==3.5.0
pytest-parallel==0.1.1
pytest-rerunfailures==12.0

# Web Automation
selenium==4.15.0
playwright==1.40.0
webdriver-manager==4.0.1

# API Testing
requests==2.31.0
pytest-postman==0.1.0
python-dotenv==1.0.0

# Reporting & Logging
allure-pytest==2.13.2
pytest-html-reporter==0.2.9
loguru==0.7.2

# Data Handling
PyYAML==6.0.1
pandas==2.1.3
openpyxl==3.1.2

# Utilities
pillow==10.1.0
faker==20.1.0
jsonschema==4.20.0
python-dateutil==2.8.2

# Code Quality
pylint==3.0.3
black==23.11.0